Include information about the window's internal line counter
I'm not exactly sure on where to put this information in, so I just dropped it here, It would be good to describe an example too, maybe also a little diagram explaining it? I can do that if you want me to, but I decided to just keep it like this until I get some more info about PRs on this doc (such as keeping the design language of diagrams consistent).