Change "wavelength" nomenclature to "period value"
Long ago, the name of the 11-bit value in NR13-NR14, NR23-NR24,
and NR33-NR34 was changed from "frequency", which is wrong, to
"wavelength", which is less wrong.  It stuck because nobody
objected (consensus by crickets).  Some time later, jvsTSX
complained that "wavelength" is confusing because the word is
more often understood as the reciprocal of spatial frequency,
not temporal frequency.  It was pointed out that documents for
other PSGs using a frequency divider also use "period" as the
name of the parameter.

Add some worked examples to help readers understand the relationship
among period value, period, sample rate, and tone frequency.