Add a SVG to visually describe the layout of VRAM
It is slightly dynamic, relying on a fuckton of pre-generated elements
and a `display:` CSS rule. This generates a **monster** SVG (5108 lines),
which is not great.

The file could be slimmed down tremendously by making use of `<use>` elements
(e.g. `<use href="#tile_block">` in six places), but this is incompatible
with the CSS trick used (then the `<text>` elements cannot be all positioned
in the same place in the parent).
It may be more reasonable to use some JS for this, instead of this pure-CSS hack.