Memory Map: add VRAM interactive figure (#544)
* Move IO map to just below general memory map

This makes it more prominent, and also more useful

* Fix up internal references within `Memory Map`

* Add a SVG to visually describe the layout of VRAM

It is slightly dynamic, relying on a fuckton of pre-generated elements
and a `display:` CSS rule. This generates a **monster** SVG (5108 lines),
which is not great.

The file could be slimmed down tremendously by making use of `<use>` elements
(e.g. `<use href="#tile_block">` in six places), but this is incompatible
with the CSS trick used (then the `<text>` elements cannot be all positioned
in the same place in the parent).
It may be more reasonable to use some JS for this, instead of this pure-CSS hack.

* Massively reduce VRAM map file size using a bit of JS

The SVG went from 929744 to 9249 bytes!!
Meanwhile, we only have 29 sloc of JS in ~3 KiB, which perform very
little processing.
Definitely a worthwhile tradeoff.

The SVG being significantly deflated also makes it easy to process by hand,
removing the need for a generator script and simplifying the build.

* Move labels for bank 1 to its right

* Add a description of the VRAM memory map

* Calculate mapAddr in a separate line

Co-authored-by: Sylvie <35663410+Rangi42@users.noreply.github.com>

* Check all properties of hover info objects

Co-authored-by: Rangi42 <remy.oukaour+rangi42@gmail.com>

* Make the SVG fully interactive stand-alone as well

* Polish up the SVG

Co-authored-by: Sylvie <35663410+Rangi42@users.noreply.github.com>

* Be a little more concise when linking to "blocks" explanation

Co-authored-by: Antonio Vivace <avivace4@gmail.com>

---------

Co-authored-by: Sylvie <35663410+Rangi42@users.noreply.github.com>
Co-authored-by: Rangi42 <remy.oukaour+rangi42@gmail.com>
Co-authored-by: Antonio Vivace <avivace4@gmail.com>